delete notes, persist storage
i was having problems persisting storage and getting correct values after some of the notes were deleted. Instead of storing the data in a separate array i'm awaiting the response and updating it. I'm sending a message with the data from popup.js with the updated local storage and pushing it back into the original array so that storage always has the correct amount of notes