added a name key to the array
i added a name property to the array where i'm storing data. The idea is to create a category for each date that the user adds a note on. I can acces the name and display different categories based on that date. If each entry has the same date it will be added to a new category, if not a new entry will be created with a new date