set active category, add to active category
i decided to improve the code so that the user can specify which category the notes will be added to. If the user wants to select a specific category where the notes should be added he just has to click that category and it will be set to active. By default all of the notes are pushed to the current date category. I'm rerendering HTML every time the user selects the category so that only the active category notes are being displayed in the popup