added a way to rename each category
each of the categories can be renamed and the new name will be stored. In order to track which item is gets a new name i'm using a rename property that is set to false by default. If the rename button  is clicked the rename property is changed to active and the form appears that takes user input for a new  name